comparison of integer expressions of different signedness
comparison of integer expressions of different signedness: ‘int’ and
‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

+++ src/kdirtreecache.cpp
@@ -179,7 +179,7 @@ static void setStateRecursive(KDirInfo * root) {
   // Once the whole cache file is read set the state of each KDirInfo to
   // finished so the tree view can start fetching the tree
   root->setReadState(KDirFinished);
-  for(int i = 0; i < root->numChildren(); i++) {
+  for(size_t i = 0; i < root->numChildren(); i++) {
     KFileInfo * f = root->child(i);
     if(f->isDirInfo()) {
       setStateRecursive(static_cast<KDirInfo*>(f));
